Second functional commit.
This development commit tidies up the project file structure a bit, and
takes the "irt" script to a basic level of functionality, where you can
use it to generate a report based off of a given CSV population file and
a specific YAML schema file; there are a number of command-line switches
to control the output, just run "scripts/irt --help" for usage information.
Finally, the Operations YAML schema file is fleshed out enough to handle
the September 2016 IDoneThis data for the Operations team.

This commit will serve as the basis for the v.0.1.0 release.