Third functional commit.
This commit includes tweaks and modifications made to the IReportedThis code
through the end of October 2016, and includes a number of small changes to the
Operation team schema YAML, plus the addition of a "tag filtering" function to
allow cleanup of individual tags as part of the reporting process, plus some
small modifications to the report output.

This commit also adds the "exsch" ("extract schema") utility, which can be used to
pretty-print more readable versions of the YAML schemas used in IReportedThis.
At the moment, it offers a Markdown-friendly format, or a "tags only" format that
just prints the hashtags for the schema, both in outline format.

The pretty-printing is actually done inside the YAMLSchema class, since it relies
on the internal data representation of the schema tree -- there's probably a
cleaner way to do this, but that's for down the road if the pretty-printing grows
in complexity significantly.

This commit will serve as the basis for the v.0.2.0 release.